requirements.txt
# Core Dependencies
streamlit>=1.28.0
streamlit-extras>=0.3.0

# ML and NLP
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.35.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# File Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
chardet>=5.2.0

# Utilities
loguru>=0.7.0
python-dotenv>=1.0.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0